Add promise support to get method
To get the results and the thrown errors, get method was
only accepting the success and error parameters respectively.
However, Dashiki uses jQuery promises returned by the
apis to subscribe callbacks.
To adapt to Dashiki and to implement a common interface,
this fix adds promise support for get method.

Note that now both options are supported,
success and error callbacks, and also callback subscription
on the returned promise.

Bug: 68448
Change-Id: I719960b677b66f60fbe2c28e8028b1a220b5567e